create and customize default_activity in activities
When a default_activity is created or updated, the listener creates or updates the activities for each user.

api customize-default allows to customize default activities.

wasCustomized field indicates if it has been customized by the user, if so, when the default is updated by an admin, the user is not affected, only the ones that have not been customized change.